Feat: Like feature add
좋아요 기능 추가 새로고침 해야 화면에 반영되는 현상은 수정예정

+++ donghang/src/main/java/com/tomato/donghang/model/service/EduApplyBoardServiceImpl.java
@@ -403,26 +403,42 @@ public class EduApplyBoardServiceImpl implements EduApplyBoardService {
 	}
 
 	@Override
-	public Integer likeCount(String userId, long eduBoardNo) {
+	public Integer likeCount(long eduBoardNo) {
 		Map<String, Object> map = new HashMap<>();
-		map.put("userId", userId);
 		map.put("eduBoardNo", eduBoardNo);
+		log.info("여기는 로그인안했을때eduApplyBoardMapper.isLike(map) {}", eduApplyBoardMapper.isLike(map));
 		return eduApplyBoardMapper.isLike(map);
 	}
-	@Override
-	public void addLikeCount(String userId, long eduBoardNo) {
-		Map<String, Object> map = new HashMap<>();
-		map.put("userId", userId);
-		map.put("eduBoardNo", eduBoardNo);
-		eduApplyBoardMapper.addLikeCount(map);
-	}
+//	@Override
+//	public void addLikeCount(String userId, long eduBoardNo) {
+//		Map<String, Object> map = new HashMap<>();
+//		map.put("userId", userId);
+//		map.put("eduBoardNo", eduBoardNo);
+//		eduApplyBoardMapper.addLikeCount(map);
+//	}
+//
+//	@Override
+//	public void deleteLikeCount(String userId, long eduBoardNo) {
+//		Map<String, Object> map = new HashMap<>();
+//		map.put("userId", userId);
+//		map.put("eduBoardNo", eduBoardNo);
+//		eduApplyBoardMapper.deleteLikeCount(map);
+//	}
 
 	@Override
-	public void deleteLikeCount(String userId, long eduBoardNo) {
+	public void likeCaculate(String userId, String value) {
 		Map<String, Object> map = new HashMap<>();
+		long eduBoardNo = Long.parseLong(value);
 		map.put("userId", userId);
 		map.put("eduBoardNo", eduBoardNo);
-		eduApplyBoardMapper.deleteLikeCount(map);
+		log.info("map {}", map);
+		Integer likeCount = eduApplyBoardMapper.isLike(map);
+		log.info("likeCount {}",likeCount);
+		if(likeCount==null || likeCount==0) {
+			eduApplyBoardMapper.addLikeCount(map);
+		}else {
+			eduApplyBoardMapper.deleteLikeCount(map);
+		}
 	}
 	
 }
